tests/wpt/tests/css/css-toggle/parsing/toggle-trigger-computed.tentative.html

<!DOCTYPE HTML>
<meta charset="UTF-8">
<title>CSS Toggles: getComputedStyle() for toggle-trigger property</title>
<link rel="author" title="L. David Baron" href="https://dbaron.org/">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://tabatkins.github.io/css-toggle/#toggle-trigger-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<body>
<div id="target"></div>
<script>

test_computed_value('toggle-trigger', 'none');
test_computed_value('toggle-trigger', 'self set 1');
test_computed_value('toggle-trigger', 'sticky set 1');
test_computed_value('toggle-trigger', 'group set 1');
test_computed_value('toggle-trigger', 'mytoggle');
test_computed_value('toggle-trigger', 'mytoggle set 0');
test_computed_value('toggle-trigger', 'mytoggle set +0', 'mytoggle set 0');
test_computed_value('toggle-trigger', 'mytoggle set -0', 'mytoggle set 0');
test_computed_value('toggle-trigger', 'mytoggle set +5', 'mytoggle set 5');
test_computed_value('toggle-trigger', 'mytoggle, yourtoggle');
test_computed_value('toggle-trigger', 'mytoggle, mytoggle');
test_computed_value('toggle-trigger', 'mytoggle set 1, yourtoggle');
test_computed_value('toggle-trigger', 'mytoggle set 1 , yourtoggle set 1', 'mytoggle set 1, yourtoggle set 1');
test_computed_value('toggle-trigger', 'mytoggle,yourtoggle set 1', 'mytoggle, yourtoggle set 1');
test_computed_value('toggle-trigger', 'mytoggle set calc(-3)', 'mytoggle set 0');
test_computed_value('toggle-trigger', 'mytoggle set calc(6)', 'mytoggle set 6');
test_computed_value('toggle-trigger', 'mytoggle set calc(6.4)', 'mytoggle set 6');
test_computed_value('toggle-trigger', 'mytoggle set calc(6.5)', 'mytoggle set 7');
test_computed_value('toggle-trigger', 'mytoggle set calc(6.6)', 'mytoggle set 7');
test_computed_value('toggle-trigger', 'mytoggle set two');
test_computed_value('toggle-trigger', 'mytoggle next 1', 'mytoggle');
test_computed_value('toggle-trigger', 'mytoggle next 2');
test_computed_value('toggle-trigger', 'mytoggle prev 1', 'mytoggle prev');
test_computed_value('toggle-trigger', 'mytoggle prev 2');
test_computed_value('toggle-trigger', 'mytoggle next calc(-3)', 'mytoggle');
test_computed_value('toggle-trigger', 'mytoggle prev calc(-3)', 'mytoggle prev');
test_computed_value('toggle-trigger', 'mytoggle next calc(6.4)', 'mytoggle next 6');
test_computed_value('toggle-trigger', 'mytoggle next calc(6.5)', 'mytoggle next 7');
test_computed_value('toggle-trigger', 'mytoggle prev calc(6.4)', 'mytoggle prev 6');
test_computed_value('toggle-trigger', 'mytoggle prev calc(6.5)', 'mytoggle prev 7');

</script>
